Oprava chyb v prezentační vrstvě
Byly opraveny určité chyby a nedostatky v prezentační vrstvě - testovací podotázky nyní mají obrázek, uživatelé jsou přesměrováni na správnou podotázku po jejím přidání/odebrání, byla opravena chyba související s přidáváním možných a správných odpovědí s diakritikou a všechna čísla by již měla být zaokrouhlena na 2 desetinná čísla.

+++ EduEnhancer/DataLayer/DataFunctions.cs
@@ -368,8 +368,12 @@ namespace DataLayer
 
         public void DeleteSubquestionTemplateImage(string webRootPath, string imageSource)
         {
-            string uploadsFolder = Path.Combine(webRootPath, "Uploads/");
-            File.Delete(uploadsFolder + imageSource);
+            //TestingImage doesn't get deleted (all testing data subquestion templates share this image)
+            if(imageSource != "TestingImage.png")
+            {
+                string uploadsFolder = Path.Combine(webRootPath, "Uploads/");
+                File.Delete(uploadsFolder + imageSource);
+            }
         }
 
         public async Task AddSubquestionTemplateStatistics(SubquestionTemplateStatistics subquestionTemplateStatistics)
